Adding support for colon and apostrophe.
Jim had some great additions including a trick to support displays that
don't have apos/colon. Because colons and apos are their own LEDs, I
moved the apos/colon handling outside the main digit loop. I treat the
apos/colon as LEDs rather than segments on a digit. This made it easier
for me to understand what was/was not working.