Support for individual segment control
If a character-to-display is  between 128-255 (the msb is set) the
display goes into bit-per-segment mode. In this case, the lower  7 bits
are each mapped to a segment.
A change like this will be required to enable the [individual segment
control](https://github.com/sparkfun/Serial7SegmentDisplay/wiki/Special-Commands#wiki-individual)
command on the S7S.